refactor: Proper task for muzzle printReferences
