+++ buildSrc/src/main/kotlin/datadog/gradle/plugin/muzzle/tasks/AbstractMuzzleTask.kt
@@ -1,0 +1,9 @@
+package datadog.gradle.plugin.muzzle.tasks
+
+import org.gradle.api.DefaultTask
+
+abstract class AbstractMuzzleTask : DefaultTask() {
+  init {
+    group = "Muzzle"
+  }
+}
